requirements.txt
email-validator==2.3.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
python-dotenv==0.21.0
pillow==10.3.0
